ci-monitoring/mcp/watcher/pyproject.toml
[project]
name = "watcher-mcp"
version = "0.1.0"
description = "MCP server for HCP/OPP Watcher triage workflows"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.28.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "sippy-mcp>=0.1.0",
]

[project.scripts]
watcher-mcp = "watcher_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/watcher_mcp"]
